% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bjw_other_function.R
\name{bjw_other}
\alias{bjw_other}
\title{Calculate the Just World Scale for Others}
\usage{
bjw_other(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
and calculates a Belief in a Just World for others score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:7)
df <- data.frame(matrix(sample(x, 10*8, replace = TRUE), nrow = 5, ncol = 8))
bjw_other(1, 8, df)
}
\references{
Lipkus, I. M., & Bissonnette, V. L. (1996). Relationships among belief in a just world, willingness to accommodate, and marital well-being. Personality and Social Psychology Bulletin, 22(10), 1043-1056.
}
\keyword{Belief}
\keyword{Just}
\keyword{Scales}
\keyword{World}
\keyword{World,}
\keyword{a}
\keyword{in}
