% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_power_function.R
\name{chr_power}
\alias{chr_power}
\title{Calculate the Generalise Sense of Power Scale}
\usage{
chr_power(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
recodes requiste columns; and calculates the sense of power score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:7)
df <- data.frame(matrix(sample(x, 10*8, replace = TRUE), nrow = 5, ncol = 8))
chr_power(1, 8, df)
}
\references{
Anderson, C., John, O. P., & Keltner, D. (2012). The personal sense of power. Journal of Personality, 80(2), 313-344.
}
\keyword{Generalise}
\keyword{Power}
\keyword{Power,}
\keyword{Scale}
\keyword{Sense}
\keyword{The}
\keyword{of}
