% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_function.R
\name{ci}
\alias{ci}
\title{Standard Confidence Interval Calculation}
\usage{
ci(x, y = 0.95, na.rm = FALSE)
}
\arguments{
\item{x}{An R object. Accepts three forms of the extract operator, the slot operator is untested.}

\item{y}{Define the confidence interval. Defaults to 0.95.}

\item{na.rm}{a logical value indicating whether NA values should be removed before computation. Defaults to FALSE.}
}
\description{
Reports the mean, confidence interval, lower confidence interval, and upper confidence interval for a range of scores.
}
\examples{
x <- sample(1:100, size = 100, replace = TRUE)
ci(x)
}
\keyword{Confidence}
\keyword{Interval}
