% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_table_function.R
\name{cor_table}
\alias{cor_table}
\title{Correlation Table with Significance Stars}
\usage{
cor_table(x, method = c("pearson", "spearman"), removeTriangle = c("upper",
  "lower"), result = c("none", "html", "latex"))
}
\arguments{
\item{x}{is a data frame containing the data.}

\item{method}{: correlation method. "pearson"" or "spearman"" is supported.}

\item{removeTriangle}{: remove upper or lower triangle.}

\item{result}{:  if "html" or "latex" the results will be displayed in html or latex format.}
}
\description{
Generates a correlation table with significance stars.
}
\examples{
x <- data.frame(replicate(5, rnorm(5)))
cor_table(x)
}
\keyword{Correlation}
\keyword{Table}
