% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dass_function.R
\name{dass}
\alias{dass}
\title{Calculate the Depression, Anxiety, and Stress Scale (DASS-21)}
\usage{
dass(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
and calculates domain scores for depression, anxiety, and stress.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(0:3)
df <- data.frame(matrix(sample(x, 10*21, replace = TRUE), nrow = 5, ncol = 21))
dass(1, 21, df)
}
\references{
Henry, J. D., & Crawford, J. R. (2005). The short‐form version of the Depression Anxiety Stress Scales (DASS‐21): Construct validity and normative data in a large non‐clinical sample. British Journal of Clinical Psychology, 44(2), 227-239.
}
\keyword{21}
\keyword{Anxiety}
\keyword{Anxiety,}
\keyword{Depression}
\keyword{Depression,}
\keyword{Scale}
\keyword{Stress}
\keyword{Stress,}
\keyword{and}
