% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_function.R
\name{des}
\alias{des}
\title{A Simple Descriptive Statistics Function}
\usage{
des(x, na.rm = FALSE)
}
\arguments{
\item{x}{An R object. Accepts three forms of the extract operator, the slot operator is untested.}

\item{na.rm}{a logical value indicating whether NA values should be removed before computation. Defaults to FALSE.}
}
\description{
Reports the mean, standard deviation, median, and inter-quartile range. Simple, decluttered output.
}
\examples{
x <- sample(1:100, size = 100, replace = TRUE)
des(x)
}
\keyword{Descriptive}
\keyword{Statistics}
