% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_function.R
\name{fl}
\alias{fl}
\title{Calculate the Flourishing Scale}
\usage{
fl(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
and calculates an overall flourishing score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:7)
df <- data.frame(matrix(sample(x, 10*10, replace = TRUE), nrow = 5, ncol = 10))
fl(1, 10, df)
}
\references{
Diener, E., Wirtz, D., Tov, W., Kim-Prieto, C., Choi, D. W., Oishi, S., & Biswas-Diener, R. (2010). New well-being measures: Short scales to assess flourishing and positive and negative feelings. Social Indicators Research, 97(2), 143-156.
}
\keyword{Flourishing}
\keyword{Flourishing,}
\keyword{Scale}
\keyword{The}
\keyword{Wellbeing,}
