% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_sat_function.R
\name{lf_sat}
\alias{lf_sat}
\title{Calculate the Satisfaction with Life Scale}
\usage{
lf_sat(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
and calculates an overall satisfaction with life score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:7)
df <- data.frame(matrix(sample(x, 10*5, replace = TRUE), nrow = 5, ncol = 5))
lf_sat(1, 5, df)
}
\references{
Diener, E. D., Emmons, R. A., Larsen, R. J., & Griffin, S. (1985). The satisfaction with life scale. Journal of Personality Assessment, 49(1), 71-75.
}
\keyword{Life}
\keyword{Life,}
\keyword{Satisfaction}
\keyword{Scale}
\keyword{The}
\keyword{Wellbeing,}
\keyword{with}
