% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlq_function.R
\name{mlq}
\alias{mlq}
\title{Calculate the Meaning in Life Questionnaire}
\usage{
mlq(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
recodes requiste columns; and calculates the domains of presence and search for meaning in life.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:7)
df <- data.frame(matrix(sample(x, 10*10, replace = TRUE), nrow = 5, ncol = 10))
mlq(1, 10, df)
}
\references{
Steger, M. F., Frazier, P., Oishi, S., & Kaler, M. (2006). The meaning in life questionnaire: Assessing the presence of and search for meaning in life. Journal of Counseling Psychology, 53(1), 80-93.
}
\keyword{Life}
\keyword{Life,}
\keyword{Meaning}
\keyword{Quesitionnaire}
\keyword{The}
\keyword{in}
