% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_check_function.R
\name{norm_check}
\alias{norm_check}
\title{Normality Test}
\usage{
norm_check(x)
}
\arguments{
\item{x}{An R object. Accepts three forms of the extract operator, the slot operator is untested.}
}
\description{
Produce a distribution plot including a density curve, mean (red line), and median (blue line).
A Shapiro-Wilk normality test is also run and results printed to console. Remember, if the Shapiro-Wilk test is significant the
data is non-normally distributed.
}
\examples{
x <- sample(1:100, size = 1000, replace = TRUE)
norm_check(x)
}
\keyword{Distribution}
\keyword{Normal}
