% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_function.R
\name{opt}
\alias{opt}
\title{Calculate the Life Orientation Test - Revised (LOT-R)}
\usage{
opt(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
recodes requiste columns; and calculates a LOT-R score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(0:4)
df <- data.frame(matrix(sample(x, 10*10, replace = TRUE), nrow = 5, ncol = 10))
opt(1, 10, df)
}
\references{
Scheier, M. F., Carver, C. S., & Bridges, M. W. (1994). Distinguishing optimism from neuroticism (and trait anxiety, self-mastery, and self-esteem): A reevaluation of the Life Orientation Test. Journal of Personality and Social Psychology, 67(6), 1063-1078.
}
\keyword{-}
\keyword{Life}
\keyword{Optimism,}
\keyword{Orientation}
\keyword{Revised}
\keyword{Test}
