% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_function.R
\name{perma}
\alias{perma}
\title{Calculate the PERMA Profiler}
\usage{
perma(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
and calculates domain scores for Positive emotion, Engagement, Relationships, Meaning, Accomplishment, General health,
Negative emotion, and overall wellbeing.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(0:10)
df <- data.frame(matrix(sample(x, 10*23, replace = TRUE), nrow = 5, ncol = 23))
perma(1, 23, df)
}
\references{
Butler, J., & Kern, M. L. (2016). The PERMA-Profiler: A brief multidimensional measure of flourishing. International Journal of Wellbeing, 6(3), 1-48.
}
\keyword{PERMA}
\keyword{Profiler}
\keyword{Wellbeing,}
