% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy_well_function.R
\name{psy_well}
\alias{psy_well}
\title{Calculate the Psychological Wellbeing Scale}
\usage{
psy_well(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
recodes requiste columns; and calculates domain scores for Autonomy, Environmental Mastery, Personal Growth,
Positive Relations, Purpose in Life, Self-acceptance, and Overall wellbeing.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:6)
df <- data.frame(matrix(sample(x, 10*42, replace = TRUE), nrow = 5, ncol = 42))
psy_well(1, 42, df)
}
\references{
Ryff, C. D., & Keyes, C. L. M. (1995). The structure of psychological well-being revisited. Journal of Personality and Social Psychology, 69(4), 719.
}
\keyword{Psychological}
\keyword{Scale}
\keyword{Wellbeing}
\keyword{Wellbeing,}
