% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci_function.R
\name{rci}
\alias{rci}
\title{Calculate the Religious Commitment Inventory}
\usage{
rci(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
and calculates an Intrapersonal Religious Commitment score, an Interpersonal Commitment score, and an overall Religious
Commitment score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:5)
df <- data.frame(matrix(sample(x, 10*10, replace = TRUE), nrow = 5, ncol = 10))
rci(1, 10, df)
}
\references{
Worthington Jr, E. L., Wade, N. G., Hight, T. L., Ripley, J. S., McCullough, M. E., Berry, J. W., ... & O'connor, L. (2003). The Religious Commitment Inventory-10: Development, refinement, and validation of a brief scale for research and counseling. Journal of Counseling Psychology, 50(1), 84-96.
}
\keyword{Commitment}
\keyword{Inventory}
\keyword{Religion,}
\keyword{Religious}
\keyword{The}
