% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_function.R
\name{re}
\alias{re}
\title{Calculate the Brief Resilience Scale}
\usage{
re(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
recodes requiste columns; and calculates a Brief Resilience Scale score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(1:5)
df <- data.frame(matrix(sample(x, 10*6, replace = TRUE), nrow = 5, ncol = 6))
re(1, 6, df)
}
\references{
Smith, B. W., Dalen, J., Wiggins, K., Tooley, E., Christopher, P., & Bernard, J. (2008). The brief resilience scale: assessing the ability to bounce back. International Journal of Behavioral Medicine, 15(3), 194-200.
}
\keyword{Brief}
\keyword{Resilience}
\keyword{Resilience,}
\keyword{Scale}
