% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_10point_function.R
\name{re_10point}
\alias{re_10point}
\title{Calculate the Brief Resilience Scale with a 10-point response scale}
\usage{
re_10point(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
recodes requiste columns; and calculates a Brief Resilience Scale score on a 10-point scale. The question order should be
the three positively worded questions followed by the three negatively worded questions.
}
\note{
This function is designed to work with a question order where the first three questions are positive and the second
three questions are negative. This function will give inaccurate results if question order is different.
}
\examples{
x <- c(0:10)
df <- data.frame(matrix(sample(x, 10*6, replace = TRUE), nrow = 5, ncol = 6))
re_10point(1, 6, df)
}
\references{
The Brief Resilience Scale as used by the SAHMRI Wellbeing and Resilience Centre.
}
\keyword{10-point}
\keyword{Brief}
\keyword{Resilience}
\keyword{Scale}
