% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr_function.R
\name{spr}
\alias{spr}
\title{Calculate the Intrinsic Spirituality Scale}
\usage{
spr(start_col, end_col, data)
}
\arguments{
\item{start_col}{The column number where the scale begins, reference by number only.}

\item{end_col}{The column number where the scale end, reference by number only.}

\item{data}{The reference dataframe.}
}
\description{
This function renames columns; converts all columns to numerics; tests if the scores are outside of the scale limits;
and calculates an overall spirituality score.
}
\note{
This function is designed to work with the validated question order as printed in the reference article. This function will
give inaccurate results if question order is different from the published validated scale.
}
\examples{
x <- c(0:10)
df <- data.frame(matrix(sample(x, 10*6, replace = TRUE), nrow = 5, ncol = 6))
spr(1, 6, df)
}
\references{
Hodge, D. R. (2003). The intrinsic spirituality scale: A new six-item instrument for assessing the salience of spirituality as a motivational construct. Journal of Social Service Research, 30(1), 41-61.
}
\keyword{Intrinsic}
\keyword{Scale}
\keyword{Spirituality}
\keyword{Spirituality,}
\keyword{The}
