% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_numeric_function.R
\name{to_numeric}
\alias{to_numeric}
\title{Convert Object to a Numeric}
\usage{
to_numeric(x)
}
\arguments{
\item{x}{An R object. Accepts three forms of the extract operator, the slot operator is untested.}
}
\description{
Robustly converts objects to numerics, converts factors to numeric as number dicplayed not reference category.
}
\examples{
x <- sample(1:100, size = 100, replace = TRUE)
to_numeric(x)
}
\keyword{numeric}
\keyword{transform}
